Move services classes to separate files
