+++ Sources/VGSCheckoutSDK/Core/Configuration/PaymentOrchestration/AddCard/VGSCheckoutAddCardConfiguration.swift
@@ -25,6 +25,9 @@ public struct VGSCheckoutAddCardConfiguration: VGSCheckoutBasicConfigurationProt
 	/// `String` object, organization vault environment with data region.(e.g. "live", "live-eu1", "sandbox"). Default is `sandbox`.
 	public let environment: String
 
+	/// Enable save card option. If enabled - button with option to save card for future payments will be displayed. Default is `true`. Default **save card button** state is `selected`.
+	public var saveCardOptionEnabled: Bool = true
+
 	/// Payment orchestration access token.
 	private(set) internal var accessToken: String
